[Maint] sync tests with refactor
